requirements.txt
flask==3.0.0
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.0
python-dotenv==1.0.0
flask_cors==4.0.0
Flask-Migrate==4.0.5
marshmallow==3.20.1
bcrypt==4.1.2
Flask-JWT-Extended==4.6.0